test(term-rewriting): own math.run input-parsing boundaries in the dispatch lane
